Adjusted exports + new Snake move method
Most function didn't need to be exported across the package since they
won't be used outside of their on file. This commit changes this since
basically everything was exported before.

Also the new move method for the snake now returns the old tail. Before,
there was an issue with a snake of length 4 that would die when going
in circle since we looked for collision before poping the snake's tail.
Now move(s::Snake) returns the old tail's position to avoid the bug
mentionned above. The old tail will be added back to the snake only when
it eats the food.